Average Calculations
Added methods calculateCurrMark() which calculates the mark based on
grades up to that point, and the calculateOverallMark() which calculates
the overall mark based on all marks within the array. Also added
capability to find a Grade object, given a String corresponding to the
name of the Grade you are looking for.